Housekeeping of the mcp project
