+++ mcp/data_processing/build_embeddings/.gitignore
@@ -1,0 +1,1 @@
+.fastembed_cache
