+++ mcp/data_processing/build_embeddings/Cargo.toml
