+++ mcp/data_processing/build_embeddings/src/main.rs
