+++ mcp/data_processing/build_examples/Cargo.toml
