+++ mcp/data_processing/pdf_parser/README.md
