+++ mcp/data_processing/pdf_parser/src/main.rs
