+++ mcp/data_processing/text_processor/README.md
