+++ mcp/data_processing/text_processor/src/bullet_fixer.rs
