+++ mcp/data_processing/text_processor/src/cleanup.rs
