+++ mcp/data_processing/text_processor/src/glyph_widths.rs
