--- mcp/guides/.gitkeep
@@ -1,2 +1,0 @@
-# This directory contains extracted text from PDF files
-# Generated by pdf_parser tool
