+++ mcp/server/src/lib.rs
