+++ mcp/server/src/main.rs
