+++ mcp/server/src/model.rs
